+++ PackageRegistry/Core/CredentialManager.cs.meta
