+++ PackageRegistry/Core/RegistryManagerController.cs.meta
