+++ PackageRegistry/Core/ScopedRegistry.cs.meta
