+++ PackageRegistry/Publishing.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: db6cc9c27300765c7bd3d196b193741f
+guid: d3d3fd7f718308a45a70f7e9b6a8dc77
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
