+++ PackageRegistry/Publishing/PublicationManifest.cs.meta
