+++ PackageRegistry/UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a67360ebdc1e6d247b183859bd4be8a1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
