+++ PackageRegistry/UI/ScopedRegistryEditorView.cs
@@ -1,9 +1,10 @@
 
 using System;
+using Halodi.PackageRegistry.Core;
 using UnityEditor;
 using UnityEngine;
 
-namespace Halodi.PackageRegistry
+namespace Halodi.PackageRegistry.UI
 {
     class ScopedRegistryEditorView : EditorWindow
     {
