+++ PackageRegistry/com.halodi.halodi-unity-package-registry-manager.Editor.asmdef.meta
